backend/app/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
langgraph
dateparser